fix: call `register_coders!()` once in all test cases
